Move hubbard model
