+++ qualtran/bloqs/phase_estimation/qubitization_qpe.py
@@ -161,7 +161,9 @@ class QubitizationQPE(GateWithRegisters):
 def _qubitization_qpe_hubbard_model_small() -> QubitizationQPE:
     import numpy as np
 
-    from qualtran.bloqs.hubbard_model import get_walk_operator_for_hubbard_model
+    from qualtran.bloqs.chemistry.hubbard_model.qubitization import (
+        get_walk_operator_for_hubbard_model,
+    )
     from qualtran.bloqs.phase_estimation import QubitizationQPE
 
     x_dim, y_dim, t = 2, 2, 2
@@ -182,7 +184,9 @@ def _qubitization_qpe_hubbard_model_small() -> QubitizationQPE:
 def _qubitization_qpe_hubbard_model_large() -> QubitizationQPE:
     import numpy as np
 
-    from qualtran.bloqs.hubbard_model import get_walk_operator_for_hubbard_model
+    from qualtran.bloqs.chemistry.hubbard_model.qubitization import (
+        get_walk_operator_for_hubbard_model,
+    )
     from qualtran.bloqs.phase_estimation import QubitizationQPE
 
     x_dim, y_dim, t = 20, 20, 20
